server/requirements.txt
authlib==0.15.5
Flask==1.1.2
flask_wtf==1.0.0
matplotlib==3.5.0
numpy==1.21.2
pandas==1.3.4
gunicorn==20.1.0
scikit_learn==1.0.1
git+https://github.com/maxcountryman/flask-uploads.git@f66d7dc
WTForms==3.0.0
python-dotenv==0.19.2
requests==2.26.0